refactor(WebAPI): Move design-time app build info to separate class
